+++ app/src/main/java/com/uzlov/valitova/justcargo/repo/repositories/RequestRepositoryImpl.kt
@@ -10,13 +10,14 @@ class RequestRepositoryImpl @Inject constructor(var remoteDataSource: IRequestsR
     IRequestsRepository {
     override fun getRequests(): LiveData<List<Request>> = remoteDataSource.getRequests()
 
-    override fun getRequest(id: Int): LiveData<Request?> = remoteDataSource.getRequest(id)
-    override fun getRequestsWithStatus(id: Int): LiveData<List<Request>> =remoteDataSource.getRequestsWithStatus(id)
+    override fun getRequest(id: Long): LiveData<Request?> = remoteDataSource.getRequest(id)
+    override fun getRequestsWithStatus(id: Int): LiveData<List<Request>> =
+        remoteDataSource.getRequestsWithStatus(id)
     override fun getRequestsWithPhone(phone: String) = remoteDataSource.getRequestsWithPhone(phone)
 
     override fun getRequestsWithUserID(id: Int): LiveData<List<Request>> = remoteDataSource.getRequestsWithUserID(id)
 
-    override fun removeRequest(id: Int) = remoteDataSource.removeRequest(id)
+    override fun removeRequest(id: Long) = remoteDataSource.removeRequest(id)
 
     override fun putRequest(request: Request) = remoteDataSource.putRequest(request)
 }
